# app/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Sadko Archive REST API.",
        "title": "Sadko Archive API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/create_content_state": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentState"
                ],
                "summary": "Create content state",
                "parameters": [
                    {
                        "description": "Only title field is accepted, the rest are ignored",
                        "name": "ContentState",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ContentState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/create_record": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Create record",
                "parameters": [
                    {
                        "description": "pcode, content_state_id are required; archived_date, last_treat are optional",
                        "name": "Record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Record"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.Record"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "integer"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/delete_content_state/{id}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentState"
                ],
                "summary": "Delete content state",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ContentState ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/get_content_states": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentState"
                ],
                "summary": "Get content states",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.ContentState"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/get_records": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Get records",
                "parameters": [
                    {
                        "description": "The fields EndCreationDate, EndArchivedDate, EndLastTreat default to the current date ",
                        "name": "RecordSearchParams",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.RecordSearchParams"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Record"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/update_content_state": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ContentState"
                ],
                "summary": "Update content state",
                "parameters": [
                    {
                        "description": "Needs id and new title to change",
                        "name": "ContentState",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.ContentState"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/update_record": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Record"
                ],
                "summary": "Update record",
                "parameters": [
                    {
                        "description": "id is required",
                        "name": "Record",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.Record"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/entity.ResponseView"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "entity.ContentState": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "entity.Record": {
            "type": "object",
            "properties": {
                "archived_date": {
                    "type": "integer"
                },
                "box_id": {
                    "type": "integer"
                },
                "branch": {
                    "type": "string"
                },
                "content_state_id": {
                    "type": "integer"
                },
                "creation_date": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "last_treat": {
                    "type": "integer"
                },
                "pcode": {
                    "type": "integer"
                }
            }
        },
        "entity.RecordSearchParams": {
            "type": "object",
            "properties": {
                "box_id": {
                    "type": "integer"
                },
                "branch": {
                    "type": "string"
                },
                "content_state_id": {
                    "type": "integer"
                },
                "end_archived_date": {
                    "type": "integer"
                },
                "end_creation_date": {
                    "type": "integer"
                },
                "end_last_treat": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "pcode": {
                    "type": "integer"
                },
                "start_archived_date": {
                    "type": "integer"
                },
                "start_creation_date": {
                    "type": "integer"
                },
                "start_last_treat": {
                    "type": "integer"
                }
            }
        },
        "entity.ResponseView": {
            "type": "object",
            "properties": {
                "data": {},
                "is_success": {
                    "type": "boolean"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}